requirements.txt
Django==1.9.7
djangorestframework==3.0.0
six==1.8.0
